+++ h3/sql/install/31-wkb.sql
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
--- complain if script is sourced in psql, rather than via CREATE EXTENSION
-\echo Use "ALTER EXTENSION h3 UPDATE TO 'unreleased'" to load this file. \quit
+--| # WKB indexing functions
 
+--@ availability: unreleased
 CREATE OR REPLACE FUNCTION
     h3_cell_to_boundary_wkb(cell h3index) RETURNS bytea
 AS 'h3' LANGUAGE C IMMUTABLE STRICT PARALLEL SAFE; COMMENT ON FUNCTION
@@ -26,19 +26,3 @@ IS 'Finds the boundary of the index, converts to EWKB.
 Splits polygons when crossing 180th meridian.
 
 This function has to return WKB since Postgres does not provide multipolygon type.';
-
--- deprecate extend flag
-DROP FUNCTION IF EXISTS h3_cell_to_boundary(cell h3index, extend_at_meridian boolean);
-CREATE OR REPLACE FUNCTION
-    h3_cell_to_boundary(cell h3index, extend_antimeridian boolean) RETURNS polygon
-AS 'h3' LANGUAGE C IMMUTABLE STRICT PARALLEL SAFE; COMMENT ON FUNCTION
-    h3_cell_to_boundary(h3index, boolean)
-IS 'DEPRECATED: Use `SET h3.extend_antimeridian TO true` instead';
-
-CREATE OR REPLACE FUNCTION
-    h3_cell_to_boundary(cell h3index) RETURNS polygon
-AS 'h3' LANGUAGE C IMMUTABLE STRICT PARALLEL SAFE; COMMENT ON FUNCTION
-    h3_cell_to_boundary(h3index)
-IS 'Finds the boundary of the index.
-
-Use `SET h3.extend_antimeridian TO true` to extend coordinates when crossing 180th meridian.';