+++ h3/src/extension.in.h
@@ -21,6 +21,7 @@
 
 void		_PG_init(void);
 extern bool h3_guc_strict;
+extern bool h3_guc_extend_antimeridian;
 
 typedef struct
 {
@@ -101,6 +102,11 @@ Datum		srf_return_h3_index_distances_from_user_fctx(PG_FUNCTION_ARGS);
 		"Function returning record called in context " \
 		"that cannot accept type record"			   \
 	)
+
+#define H3_DEPRECATION(msg)           \
+	ereport(WARNING, (			      \
+		errmsg("Deprecated: %s", msg) \
+	))
 
 #define DEBUG(msg, ...)			   \
 	ereport(NOTICE, (			   \