+++ h3/test/sql/postgis.sql
@@ -62,19 +62,17 @@ SELECT COUNT(*) = 48 FROM (
 SELECT h3_polygon_to_cells(h3_cell_to_boundary(:hexagon)::geometry::polygon, null, :resolution) = :hexagon;
 
 -- the boundary of of a non-edgecrossing index is a polygon
-SELECT GeometryType(h3_cell_to_boundary_wkb(:hexagon, true)::geometry) LIKE 'POLYGON';
+SELECT GeometryType(h3_cell_to_boundary_wkb(:hexagon)::geometry) LIKE 'POLYGON';
 
 -- the boundary of an edgecrossing index is a multipolygon when split
-SELECT GeometryType(h3_cell_to_boundary_wkb(:edgecross, true)::geometry) LIKE 'MULTIPOLYGON';
-
--- the boundary of an edgecrossing index is a polygon when not split
-SELECT GeometryType(h3_cell_to_boundary_wkb(:edgecross, false)::geometry) LIKE 'POLYGON';
+SELECT GeometryType(h3_cell_to_boundary_wkb(:edgecross)::geometry) LIKE 'MULTIPOLYGON';
 
 -- check latitude of antimeridian crossing points
+SET h3.split_antimeridian TO true;
 SELECT every(ABS(ST_Y(p) - :lat1) < :epsilon OR ABS(ST_Y(p) - :lat2) < :epsilon)
 FROM (
     SELECT (dp).geom AS p FROM (
-        (SELECT ST_DumpPoints(h3_cell_to_boundary_wkb(:edgecross, true)::geometry) AS dp)
+        (SELECT ST_DumpPoints(h3_cell_to_boundary_wkb(:edgecross)::geometry) AS dp)
     ) AS q1
 ) AS q2
 WHERE ABS(ABS(ST_X(p)) - 180) < :epsilon;